+++ lib/ui/settings/general/auto_load_switch.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nesd/ui/common/focus_on_hover.dart';
+import 'package:nesd/ui/common/settings_tile.dart';
+import 'package:nesd/ui/settings/settings.dart';
+
+class AutoLoadSwitch extends ConsumerWidget {
+  const AutoLoadSwitch({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final setting = ref.watch(
+      settingsControllerProvider.select((s) => s.autoLoad),
+    );
+    final controller = ref.read(settingsControllerProvider.notifier);
+
+    return FocusOnHover(
+      child: SwitchSettingsTile(
+        title: const Text('Load latest save state on start'),
+        value: setting,
+        onChanged: (value) => controller.autoLoad = value,
+      ),
+    );
+  }
+}